+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Added
+
+- Add `AssociationHistory` trait, model, migration, and Blade component for tracking field changes on any model.
+
 ## [5.4.0] - 2026-03-11
 
 _Stable release based on [5.4.0-rc.1]._
@@ -1311,6 +1317,7 @@ _Stable release based on [0.1.0-rc.1]._
 
 ## [0.0.0] - 2024-09-12
 
+[Unreleased]: https://https://github.com/internetguru/laravel-common/compare/staging...dev
 [5.4.0]: https://https://github.com/internetguru/laravel-common/compare/v5.3.1...v5.4.0
 [5.4.0-rc.1]: https://github.com/internetguru/laravel-common/releases/tag/v5.3.1
 [5.3.1]: https://https://github.com/internetguru/laravel-common/compare/v5.3.0...v5.3.1